requirements.txt
altair==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
blis==0.7.11 ; python_version >= "3.10" and python_version < "4.0"
branca==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
catalogue==2.0.10 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
cloudpathlib==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
confection==0.1.4 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
cymem==2.0.8 ; python_version >= "3.10" and python_version < "4.0"
folium==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.47.0 ; python_version >= "3.10" and python_version < "4.0"
geographiclib==2.0 ; python_version >= "3.10" and python_version < "4.0"
geopy==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.20.0 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
langcodes==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.8.2 ; python_version >= "3.10" and python_version < "4.0"
murmurhash==1.0.10 ; python_version >= "3.10" and python_version < "4.0"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.1.4 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.10" and python_version < "4.0"
plotly==5.18.0
preshed==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.32.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2023.12.25 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.15.2 ; python_version >= "3.10" and python_version < "4.0"
setuptools==69.0.3 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
smart-open==6.4.0 ; python_version >= "3.10" and python_version < "4.0"
spacy-legacy==3.0.12 ; python_version >= "3.10" and python_version < "4.0"
spacy-loggers==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
spacy==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
srsly==2.4.8 ; python_version >= "3.10" and python_version < "4.0"
streamlit==1.29.0
streamlit_option_menu==0.3.6
tenacity==8.2.3 ; python_version >= "3.10" and python_version < "4.0"
textblob==0.17.1 ; python_version >= "3.10" and python_version < "4.0"
thinc==8.2.2 ; python_version >= "3.10" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "4.0"
typer==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
wasabi==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
weasel==0.3.4 ; python_version >= "3.10" and python_version < "4.0"
wordcloud==1.9.3 ; python_version >= "3.10" and python_version < "4.0"
